config/requirements/dev.txt
# Доп. пакеты для разработки
-r base.txt

django-debug-toolbar
